requirements.txt
alembic==1.0.11
Babel==2.7.0
blinker==1.4
Click==7.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.5
uWSGI==2.0.18
Werkzeug==0.15.4
WTForms==2.2.1
